Some progress, need to pick right wall to remove.
I see that I have the correnct option in the list of candidates, but from
what I can see the options I choose seems to be the correct one. I should
take another look at the procedure to determine the wall to remove, and also
draw out how the grid is supposed to look, to see if I should even have
my result in the candidates in the first place.